fix bug - change logic to save command to database and publish to device - save record from sensor to database

+++ src/main/java/com/smartHome/model/RecordType/Humidity.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 @DiscriminatorValue("HUMIDITY")
 public class Humidity extends Record {
     @Column(nullable = false)
-    private Double temperature;
+    private Float temperature;
 }
